Moving some things out of util/misc
