+++ docs/reference_guides/core/index.rst
@@ -15,3 +15,4 @@
     phase
     util/index
     solvers
+    var_like_expr
