+++ idaes/core/util/tests/test_misc.py
@@ -16,7 +16,7 @@ This module contains miscalaneous utility functions for use in IDAES models.
 
 import pytest
 
-from pyomo.environ import ConcreteModel, Expression, Set, Block, Var, units
+from pyomo.environ import ConcreteModel, Set, Block, Var, units
 from pyomo.network import Port, Arc
 from pyomo.common.config import ConfigBlock
 from pyomo.core.base.units_container import UnitsError
@@ -25,7 +25,6 @@ from idaes.core.util.misc import (
     add_object_reference,
     copy_port_values,
     TagReference,
-    VarLikeExpression,
     set_param_from_config,
 )
 import idaes.logger as idaeslog
@@ -151,148 +150,6 @@ def test_tag_reference():
     assert test_tag[0].value == 1
     assert test_tag[1].value == 2
     assert test_tag.description == "y tag"
-
-
-@pytest.mark.unit
-def test_SimpleVarLikeExpression():
-    m = ConcreteModel()
-
-    # Need a Var to use in the Expression to avoid being able to set the value
-    # of a float
-    m.v = Var(initialize=42)
-
-    m.e = VarLikeExpression(expr=m.v)
-
-    assert m.e.type() is Expression
-    assert not m.e.is_indexed()
-
-    with pytest.raises(
-        TypeError,
-        match="e is an Expression and does not have a value "
-        "attribute. Use the 'value\(\)' method instead.",
-    ):
-        assert m.e.value == 42
-
-    with pytest.raises(
-        TypeError,
-        match="e is an Expression and does not have a value " "which can be set.",
-    ):
-        m.e.set_value(10)
-
-    with pytest.raises(
-        TypeError,
-        match="e is an Expression and does not have a value " "which can be set.",
-    ):
-        m.e.value = 10
-
-    with pytest.raises(
-        TypeError,
-        match="e is an Expression and can not have bounds. "
-        "Use an inequality Constraint instead.",
-    ):
-        m.e.setub(10)
-    with pytest.raises(
-        TypeError,
-        match="e is an Expression and can not have bounds. "
-        "Use an inequality Constraint instead.",
-    ):
-        m.e.setlb(0)
-    with pytest.raises(
-        TypeError,
-        match="e is an Expression and can not be fixed. "
-        "Use an equality Constraint instead.",
-    ):
-        m.e.fix(8)
-    with pytest.raises(TypeError, match="e is an Expression and can not be unfixed."):
-        m.e.unfix()
-
-    m.e.set_value(10, force=True)
-    assert m.e._expr == 10
-
-
-@pytest.mark.unit
-def test_IndexedVarLikeExpression():
-    m = ConcreteModel()
-
-    # Need a Var to use in the Expression to avoid being able to set the value
-    # of a float
-    m.v = Var(initialize=42)
-
-    m.e = VarLikeExpression([1, 2, 3, 4], expr=m.v)
-
-    assert m.e.type() is Expression
-    assert m.e.is_indexed()
-
-    with pytest.raises(
-        TypeError,
-        match="e is an Expression and can not have bounds. "
-        "Use inequality Constraints instead.",
-    ):
-        m.e.setub(10)
-    with pytest.raises(
-        TypeError,
-        match="e is an Expression and can not have bounds. "
-        "Use inequality Constraints instead.",
-    ):
-        m.e.setlb(0)
-    with pytest.raises(
-        TypeError,
-        match="e is an Expression and can not be fixed. "
-        "Use equality Constraints instead.",
-    ):
-        m.e.fix(8)
-    with pytest.raises(TypeError, match="e is an Expression and can not be unfixed."):
-        m.e.unfix()
-
-    for i in m.e:
-        with pytest.raises(
-            TypeError,
-            match=f"e\[{i}\] is an Expression and does not have"
-            f" a value attribute. Use the 'value\(\)' method "
-            "instead",
-        ):
-            assert m.e[i].value == 42
-
-        with pytest.raises(
-            TypeError,
-            match=f"e\[{i}\] is an Expression and does not "
-            f"have a value which can be set.",
-        ):
-            m.e[i].set_value(10)
-
-        with pytest.raises(
-            TypeError,
-            match="e\[{}\] is an Expression and does not have "
-            "a value which can be set.".format(i),
-        ):
-            m.e[i].value = 10
-
-        with pytest.raises(
-            TypeError,
-            match="e\[{}\] is an Expression and can not have "
-            "bounds. Use an inequality Constraint instead.".format(i),
-        ):
-            m.e[i].setub(10)
-        with pytest.raises(
-            TypeError,
-            match="e\[{}\] is an Expression and can not have "
-            "bounds. Use an inequality Constraint instead.".format(i),
-        ):
-            m.e[i].setlb(0)
-        with pytest.raises(
-            TypeError,
-            match="e\[{}\] is an Expression and can not be "
-            "fixed. Use an equality Constraint instead.".format(i),
-        ):
-            m.e[i].fix(8)
-        with pytest.raises(
-            TypeError,
-            match="e\[{}\] is an Expression and can not be " "unfixed.".format(i),
-        ):
-            m.e[i].unfix()
-
-        m.e[i].set_value(i, force=True)
-        assert m.e[i]._expr == i
 
 
 @pytest.mark.unit
